git-tfs-id: [http://tfs.bumc.bu.edu:8080/tfs/BU-MICRO-MSI]$/Daphne/Daphne-axin3;C6047

+++ Daphne/Simulation.cs
@@ -1188,7 +1188,7 @@ namespace Daphne
             }
             // 5., find the maximum force
             f_max = 0;
-            foreach (KeyValuePair<int, Pair> kvp in collisionManager.Pairs)
+            foreach (KeyValuePair<long, Pair> kvp in collisionManager.Pairs)
             {
                 if (kvp.Value.Force > f_max)
                 {
